refactor parts
