+++ model/host/db.go
@@ -483,72 +483,6 @@ func GetHostsByFromIdWithStatus(id, status, user string, limit, sortDir int) ([]
 	return hosts, nil
 }
 
-// statsByDistroPipeline returns a pipeline that will group all up hosts by distro
-// and return the count of hosts as well as how many are running tasks
-func statsByDistroPipeline() []bson.M {
-	return []bson.M{
-		{
-			"$match": bson.M{
-				StatusKey: bson.M{
-					"$in": evergreen.UphostStatus,
-				},
-			},
-		},
-		{
-			"$group": bson.M{
-				"_id": bson.M{
-					"distro": "$distro._id",
-					"status": "$" + StatusKey,
-				},
-				"count": bson.M{
-					"$sum": 1,
-				},
-				"tasks": bson.M{
-					"$addToSet": "$" + RunningTaskKey,
-				},
-			},
-		},
-		{
-			"$project": bson.M{
-				"distro":            "$_id.distro",
-				"status":            "$_id.status",
-				"count":             1,
-				"num_tasks_running": bson.M{"$size": "$tasks"},
-				"_id":               0,
-			},
-		},
-	}
-}
-
-func statsByProviderPipeline() []bson.M {
-	return []bson.M{
-		{
-			"$match": bson.M{
-				StatusKey: bson.M{
-					"$in": evergreen.UphostStatus,
-				},
-			},
-		},
-		{
-			"$group": bson.M{
-				"_id": bson.M{
-					"provider": "$" + bsonutil.GetDottedKeyName(DistroKey, distro.ProviderKey),
-				},
-				"count": bson.M{
-					"$sum": 1,
-				},
-			},
-		},
-		{
-			"$project": bson.M{
-				"provider": "$_id.provider",
-				"count":    1,
-				"_id":      0,
-			},
-		},
-	}
-}
-
 // QueryWithFullTaskPipeline returns a pipeline to match hosts and embeds the
 // task document within the host, if it's running a task
 func QueryWithFullTaskPipeline(match bson.M) []bson.M {
